refacor: move main function to main.js file
